requirements.txt
asgiref==3.11.1
brotli==1.2.0
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-phonenumber-field==8.4.0
django-tinymce==5.0.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
phonenumbers==9.0.26
pillow==12.1.1
requests==2.32.3
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
whitenoise==6.12.0
boto3==1.38.29
django-storages==1.14.6
